refactor: move footer and header to organisms's directory
